Move TransferLimitDirection and its usages from db/sqla/constants.py to common/constants.py
