+++ lib/rucio/db/sqla/migrate_repo/versions/13d4f70c66a9_introduce_transfer_limits.py
@@ -19,7 +19,7 @@ import sqlalchemy as sa
 from alembic import context
 from alembic.op import create_check_constraint, create_foreign_key, create_index, create_primary_key, create_table, drop_table
 
-from rucio.db.sqla.constants import TransferLimitDirection
+from rucio.common.constants import TransferLimitDirection
 from rucio.db.sqla.types import GUID
 
 # Alembic revision identifiers
